Add more implementation issues and security considerations
Jouni pointed out that some EAP methods didn't supply EMSK.  The
existing text wasn't clear on what to do with the dual derivations of
IMSK_MSK and IMSK_EMSK.  As a result, the handling of this behavior
was implementation defined.

Analysis showed that both hostap and FreeRADIUS did the same thing:
accidentally made IMSK_EMSK[j] == IMSK_EMSK[j - 1] in that situation.
We then update the draft to document this behhavior, to assist with
future implementations.

We also note that implementations have chosen to support EAP-TLS and
EAP-FAST-MSCHAPv2 inside of phase 2.  Other EAP methods in Phase 2 are
likely not implemented, and may therefore  work only by accident.

The security considerations section has been updated to note the
implications of this behavior.